+++ lib/src/optimised_text/optimised_text.dart
@@ -1,51 +1,57 @@
 import 'package:flutter/material.dart';
 
-import 'optimal_wrap_rich_text_for_width.dart';
+import 'optimised_text_for_width.dart';
 
-/// A widget that wraps rich text optimally.
-final class OptimalWrapRichText extends StatelessWidget {
+/// A widget that wraps text optimally.
+final class OptimisedText extends StatelessWidget {
   /// Provide the width for edge cases where LayoutBuilder cannot be used.
-  /// Such as inside SliverFillRemaining.
+  /// Such as inside SliverFillRemaining
   final double? width;
 
-  /// See equivalent property in [Text.rich].
-  final InlineSpan text;
+  /// See equivalent property in [Text]
+  final String text;
 
-  /// See equivalent property in [Text.rich].
+  /// See equivalent property in [Text]
   final TextStyle? style;
 
-  /// See equivalent property in [Text.rich].
+  /// See equivalent property in [Text]
   final StrutStyle? strutStyle;
 
-  /// See equivalent property in [Text.rich].
+  /// See equivalent property in [Text]
   final TextAlign? textAlign;
 
-  /// See equivalent property in [Text.rich].
+  /// See equivalent property in [Text]
   final TextDirection? textDirection;
 
-  /// See equivalent property in [Text.rich].
+  /// See equivalent property in [Text]
   final Locale? locale;
 
-  /// See equivalent property in [Text.rich].
+  /// See equivalent property in [Text]
   final bool? softWrap;
 
-  /// See equivalent property in [Text.rich].
+  /// See equivalent property in [Text]
   final TextScaler? textScaler;
 
-  /// See equivalent property in [Text.rich].
+  /// See equivalent property in [Text]
   final String? semanticsLabel;
 
-  /// See equivalent property in [Text.rich].
+  /// See equivalent property in [Text]
   final TextWidthBasis? textWidthBasis;
 
-  /// See equivalent property in [Text.rich].
+  /// See equivalent property in [Text]
   final TextHeightBehavior? textHeightBehavior;
 
-  /// See equivalent property in [Text.rich].
+  /// See equivalent property in [Text]
   final Color? selectionColor;
 
-  /// Creates a widget that wraps rich text optimally.
-  const OptimalWrapRichText(
+  /// See equivalent property in [Text]
+  final int? maxLines;
+
+  /// See equivalent property in [Text]
+  final TextOverflow? overflow;
+
+  /// Creates a widget that wraps text optimally.
+  const OptimisedText(
     this.text, {
     super.key,
     this.width,
@@ -60,12 +66,14 @@ final class OptimalWrapRichText extends StatelessWidget {
     this.textWidthBasis,
     this.textHeightBehavior,
     this.selectionColor,
+    this.maxLines,
+    this.overflow,
   });
 
   @override
   Widget build(BuildContext context) {
     if (width != null) {
-      return OptimalWrapRichTextForWidth(
+      return OptimisedTextForWidth(
         text,
         width: width!,
         style: style,
@@ -79,12 +87,14 @@ final class OptimalWrapRichText extends StatelessWidget {
         textWidthBasis: textWidthBasis,
         textHeightBehavior: textHeightBehavior,
         selectionColor: selectionColor,
+        maxLines: maxLines,
+        overflow: overflow,
       );
     }
 
     return LayoutBuilder(
       builder: (context, constraints) {
-        return OptimalWrapRichTextForWidth(
+        return OptimisedTextForWidth(
           text,
           width: constraints.maxWidth,
           style: style,
@@ -98,6 +108,8 @@ final class OptimalWrapRichText extends StatelessWidget {
           textWidthBasis: textWidthBasis,
           textHeightBehavior: textHeightBehavior,
           selectionColor: selectionColor,
+          maxLines: maxLines,
+          overflow: overflow,
         );
       },
     );